requirements.txt
numpy
opencv-python
mlx
Pillow
imageio